requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.3.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.59.0
fsspec==2024.6.1
idna==3.4
importlib_resources==6.5.2
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==25.0
pillow==11.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.28.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sympy==1.13.3
threadpoolctl==3.6.0
tkinterdnd2==0.4.3
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
typing_extensions==4.12.2
urllib3==1.26.13
zipp==3.23.0
